Expose isPresent method
